Move AdvanceTo and Emit to inputs mod
