+++ examples/recovery_wordcount.py
@@ -1,7 +1,7 @@
 import re
 
-from bytewax import AdvanceTo, Dataflow, Emit, parse, spawn_cluster
-from bytewax.inputs import ManualInputConfig
+from bytewax import Dataflow, parse, spawn_cluster
+from bytewax.inputs import AdvanceTo, Emit, ManualInputConfig
 from bytewax.recovery import KafkaRecoveryConfig
