+++ pytests/test_recovery.py
@@ -2,9 +2,9 @@ from threading import Event
 
 from pytest import fixture, raises
 
-from bytewax import AdvanceTo, Dataflow, Emit, run_main
+from bytewax import Dataflow, run_main
 from bytewax.recovery import SqliteRecoveryConfig
-from bytewax.inputs import ManualInputConfig
+from bytewax.inputs import AdvanceTo, Emit, ManualInputConfig
 
 RECOVERY_CONFIG_TYPES = [
     "SqliteRecoveryConfig",
